+++ include/tvm/ir/type.h
@@ -51,6 +51,7 @@
 
 #include <tvm/runtime/object.h>
 #include <tvm/node/node.h>
+#include <tvm/node/env_func.h>
 #include <tvm/node/container.h>
 #include <tvm/ir/span.h>
 #include <string>
